move make_browse_image func def and call
