+++ src/hyp3_isce2/utils.py
@@ -1,3 +1,6 @@
+from osgeo import gdal
+
+
 def utm_from_lon_lat(lon: float, lat: float) -> int:
     """Get the UTM zone epsg from a longitude and latitude.
 
@@ -31,3 +34,16 @@ def extent_from_geotransform(geotransform: tuple, x_size: int, y_size: int) -> t
         geotransform[3] + geotransform[5] * y_size,
     )
     return extent
+
+
+def make_browse_image(input_tif: str, output_png: str) -> None:
+    stats = gdal.Info(input_tif, format='json', stats=True)['stac']['raster:bands'][0]['stats']
+    gdal.Translate(
+        destName=output_png,
+        srcDS=input_tif,
+        format='png',
+        outputType=gdal.GDT_Byte,
+        width=2048,
+        strict=True,
+        scaleParams=[[stats['minimum'], stats['maximum']]],
+    )
